requirements.txt
# --- Web framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# --- Database ---
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# --- HTTP client ---
httpx>=0.27.0

# --- Validation / settings ---
pydantic>=2.7.0
pydantic-settings>=2.0.0

# --- Environment ---
python-dotenv>=1.0.0

# --- Rate limiting ---
slowapi>=0.1.9
limits>=3.6.0

# --- Retry logic ---
tenacity>=8.2.0
